管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -32,47 +32,54 @@ public class EmployeeAction extends ActionBase {
      * 一覧画面を表示する
      */
     public void index() throws ServletException, IOException {
-        // 指定されたページのデータを取得
-        int page = getPage();
+        // 管理者権限チェック
+        if (checkAdmin()) {
 
-        List<EmployeeView> employees = service.getPerPage(page);
+            // 指定されたページのデータを取得
+            int page = getPage();
 
-        // 全ての従業員データの件数を取得
-        long employeeCount = service.countAll();
+            List<EmployeeView> employees = service.getPerPage(page);
 
-        putRequestScope(AttributeConst.EMPLOYEES, employees);
-        putRequestScope(AttributeConst.EMP_COUNT, employeeCount);
-        putRequestScope(AttributeConst.PAGE, page);
-        putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE);
+            // 全ての従業員データの件数を取得
+            long employeeCount = service.countAll();
 
-        //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
-        String flush = getSessionScope(AttributeConst.FLUSH);
-        if (flush != null) {
-            putRequestScope(AttributeConst.FLUSH, flush);
-            removeSessionScope(AttributeConst.FLUSH);
+            putRequestScope(AttributeConst.EMPLOYEES, employees);
+            putRequestScope(AttributeConst.EMP_COUNT, employeeCount);
+            putRequestScope(AttributeConst.PAGE, page);
+            putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE);
+
+            //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
+            String flush = getSessionScope(AttributeConst.FLUSH);
+            if (flush != null) {
+                putRequestScope(AttributeConst.FLUSH, flush);
+                removeSessionScope(AttributeConst.FLUSH);
+            }
+
+            //一覧画面を表示
+            forward(ForwardConst.FW_EMP_INDEX);
         }
-
-        //一覧画面を表示
-        forward(ForwardConst.FW_EMP_INDEX);
     }
 
     /*
      * 新規登録画面を表示する
      */
     public void entryNew() throws ServletException, IOException {
-        putRequestScope(AttributeConst.TOKEN, getTokenId());
-        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());   // 空の従業員インスタンス
+        // 管理者権限チェック
+        if (checkAdmin()) {
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());   // 空の従業員インスタンス
 
-        // 新規登録画面に遷移
-        forward(ForwardConst.FW_EMP_NEW);
+            // 新規登録画面に遷移
+            forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /*
      * 新規登録を行う
      */
     public void create() throws ServletException, IOException {
-        // CSRF対策tokenのチェック
-        if (checkToken()) {
+        // CSRF対策token、管理者権限のチェック
+        if (checkToken() && checkAdmin()) {
             // パラメータの値を元に従業員情報のインスタンスを作成
             EmployeeView ev = new EmployeeView(
                     null,
@@ -114,44 +121,48 @@ public class EmployeeAction extends ActionBase {
      * 詳細画面を表示する
      */
     public void show() throws ServletException, IOException {
+        // 管理者権限チェック
+        if (checkAdmin()) {
+            // idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-        // idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+                // データが取得できなかった、または論理削除されている場合エラー画面へ
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
 
-        if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
-            // データが取得できなかった、または論理削除されている場合エラー画面へ
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            // リクエストスコープに従業員データを設定
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
+
+            // 詳細画面を表示
+            forward(ForwardConst.FW_EMP_SHOW);
         }
-
-        // リクエストスコープに従業員データを設定
-        putRequestScope(AttributeConst.EMPLOYEE, ev);
-
-        // 詳細画面を表示
-        forward(ForwardConst.FW_EMP_SHOW);
     }
 
     /*
      * 編集画面を表示する
      */
     public void edit() throws ServletException, IOException {
+        // 管理者権限チェック
+        if (checkAdmin()) {
+            // idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-        // idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+                // データが取得できなかった、または論理削除されている場合エラー画面へ
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
 
-        if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
-            // データが取得できなかった、または論理削除されている場合エラー画面へ
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            // リクエストスコープにTOKENと従業員データを設定
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            ev.setPassword("");
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
+
+            // 編集画面を表示
+            forward(ForwardConst.FW_EMP_EDIT);
         }
-
-        // リクエストスコープにTOKENと従業員データを設定
-        putRequestScope(AttributeConst.TOKEN, getTokenId());
-        ev.setPassword("");
-        putRequestScope(AttributeConst.EMPLOYEE, ev);
-
-        // 編集画面を表示
-        forward(ForwardConst.FW_EMP_EDIT);
     }
 
     /*
@@ -159,8 +170,8 @@ public class EmployeeAction extends ActionBase {
      */
     public void update() throws ServletException, IOException {
 
-        // tokenのチェック
-        if (checkToken()) {
+        // token、管理者権限チェック
+        if (checkToken() && checkAdmin()) {
             // パラメータの値をもとに従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -203,8 +214,8 @@ public class EmployeeAction extends ActionBase {
      */
     public void destroy() throws ServletException, IOException {
 
-        // tokenのチェック
-        if (checkToken()) {
+        // token、管理者権限チェック
+        if (checkToken() && checkAdmin()) {
             // idを条件に従業員データを論理削除
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -214,5 +225,24 @@ public class EmployeeAction extends ActionBase {
             // 一覧画面にリダイレクト
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
+    }
+
+    /*
+     * ログイン権限チェック
+     * 管理者以外はエラー画面に遷移
+     *
+     * @return true:管理者／false:管理者以外
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        // セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView)getSessionScope(AttributeConst.LOGIN_EMP);
+
+        // 管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        }
+        return true;
     }
 }
